fix: rebase 후 직원 수 관련 문제 해결
