+++ src/main/java/com/sprint/hrbank_sb6_1/dto/DepartmentSearchCond.java
@@ -2,11 +2,10 @@ package com.sprint.hrbank_sb6_1.dto;
 
 public record DepartmentSearchCond(
     String nameOrDescription,
-    Integer idAfter,
+    Long idAfter,
     String cursor,                // 🔹 커서 필드 추가
-        Integer size,
-        DepartmentSortBy sortField,
-        SortDirection sortDirection
+    Integer size,
+    DepartmentSortBy sortField,
+    SortDirection sortDirection
 ) {
-  public int safeSize() { return Math.max(size, 1); }
 }
